+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* add Dev Notes and Field Guide links to dashboard
 
 #### 2.2.7 / 2020-03-02
 * update trac link in callout for _closed_ or _reopened_ tickets on the milestone